[ALPS09641331] seninf: fix coverity
[Description]
1. fix coverity

[Test]
sanity

MTK-Commit-Id: 1016559b12831878f103d132226290fb5d153943

CR-Id: ALPS09641331
Feature: [Android Default] Camera Application Basic Functions

+++ imgsensor/src-isp8s/adaptor-subdrv-ctrl.c
@@ -4184,7 +4184,7 @@ void get_exp_line_by_scenario(struct subdrv_ctx *ctx,
 	if (ctx->s_ctx.cust_get_linetime_in_us != NULL) {
 		ctx->s_ctx.cust_get_linetime_in_us((void *) ctx,
 			ctx->current_scenario_id, &linetime_in_ns, 0);
-		shutter = 1000000000 / framerate * 10
+		shutter = (u64)1000000000 / framerate * 10
 			/ linetime_in_ns;
 	} else {
 		shutter = ctx->s_ctx.mode[scenario_id].pclk / framerate * 10
